Refactor to move data storage related files into storage package
