+++ src/test/java/seedu/address/testutil/FlashcardBuilder.java
@@ -7,7 +7,7 @@ import seedu.address.model.person.Answer;
 import seedu.address.model.person.Flashcard;
 import seedu.address.model.person.Question;
 import seedu.address.model.tag.Tag;
-import seedu.address.model.util.LocalDatabasePopulator;
+import seedu.address.storage.LocalDatabasePopulator;
 
 /**
  * A utility class to help with building Flashcard objects.
